refactor: UserController 에서 AuthController 분리
