+++ MSL-back/src/main/resources/application.yml
@@ -5,14 +5,14 @@ spring:
       path: /h2-console
 
   datasource:
-    url: jdbc:h2:mem:MLS
+    url: jdbc:h2:tcp://localhost/~/msl
     driver-class-name: org.h2.Driver
     username: root
-    password: '0000'
+    password:
 
   jpa:
     hibernate:
-      ddl-auto: update
+      ddl-auto: none
 
     generate-ddl: off
 
@@ -22,8 +22,12 @@ spring:
     open-in-view: false
 
 secret:
-  access: access-test
-  refresh: refresh-test
+  access: access-mslbacksecretkeymslbacksecretkey
+
+import:
+  key: 5161304000601633
+  secret: EZiNkyjk3eOAnikh95U24zgwsDcyzD4wYmuLr3WtnG1v9LkV1t3I3PYyIlp8TTCNNx7i9yd6Lw2JK5Ki
+
 
 logging:
   level:
